Changed folder structure to suppport mulitple cnx products
* Moved fixtures into their own folder. Fixtures for a particular
product should be in `fixtures/<product_name>.py. Fixtures that deal
with the main selenium fixtures should be in `base.py`
* Moved webview tests into their own folder `tests/webview/ui`
* Moved webview pages into their own folder `pages/webview/`
