+++ pages/webview/content.py
@@ -4,7 +4,7 @@
 
 from selenium.webdriver.common.by import By
 
-from pages.base import Base
+from pages.webview.base import Base
 
 
 class Content(Base):
